pyproject.toml
[tool.poetry]
name = "animelister"
version = "0.1.0"
description = ""
authors = ["Ben Beecher <Ben@lightmatter.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django-import-export = "^2.6.1"
django-environ = "^0.8.1"
psycopg2-binary = "^2.9.1"
django-extensions = "^3.1.3"
django-model-utils = "^4.2.0"
boto3 = "^1.19.7"
django-storages = "^1.12.3"
uvicorn = "^0.15.0"
gunicorn = "^20.1.0"
sentry-sdk = "^1.4.3"
Django = "^3.2.8"
whitenoise = {extras = ["brotli"], version = "^5.3.0"}
django-allauth = "^0.45.0"
django-htmx = "^1.6.0"
django-annoying = "^0.10.6"
django-taggit = "^1.5.1"
Pillow = "^8.4.0"
django-anymail = {extras = ["sendgrid"], version = "^8.4"}

[tool.poetry.dev-dependencies]
ipython = "^7.29.0"
ipdb = "^0.13.9"
pdbpp = "^0.10.3"
ptpython = "^3.0.20"
isort = "^5.9.3"
black = "^21.9b0"
django-debug-toolbar = "^3.2.2"
model-bakery = "^1.3.3"
coverage = "^6.1.1"
watchdog = "*"
pylint-django = "==2.4.4"
Werkzeug = "*"
bandit = "*"
honcho = "*"
flake8 = "==3.8.4"
flake8-polyfill = "==1.0.2"
unittest-xml-reporting = "^3.0.4"
prospector = "^1.5.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
